+++ src/Pim/Bundle/CatalogBundle/Manager/AttributeManager.php
@@ -154,21 +154,4 @@ class AttributeManager
 
         return $choices;
     }
-
-    /**
-     * Get the attribute group choices
-     *
-     * @return array
-     */
-    public function getAttributeGroupChoices()
-    {
-        $groups = $this->objectManager->getRepository('PimCatalogBundle:AttributeGroup')->findAllWithTranslations();
-        $choices = array();
-        foreach ($groups as $group) {
-            $choices[$group->getCode()] = $group->getLabel();
-        }
-        asort($choices);
-
-        return $choices;
-    }
 }